requirements.txt
py2neo==4.3.0
neo4j-driver==1.6.3